Extract methods and classes from CryptoUtilities. Keep one utils file.
